Fixed bodies not bouncing correctly. #14
Was because rotation on physics bodies was enabled, this meant that every time an impulse was applied the physics body rotated fast and thus bounced side ways when encountering another surface. Problem was fixed by disabling rotation on physics bodies. It can still look like the objects are being rotated by drawing the textures in the correct orientation based on current velocity etc.